Updated Quiz and GUI
Did a lot of thins! Very happy with the progress!

Renamed Quiz class QuizConsole because it is the console version of the game.
Created a new Quiz class to house all methods required for the GUI.
Updated GUI to include quiz game GUI no functionality yet for quiz so not playable yet i.e. not events for the game